refactor: centralize InternalMutation type, follow naming convention for handlers/mutations
