+++ src/shared/handlers.tsx
@@ -1,0 +1,9 @@
+import { UseMutationResult } from '@tanstack/react-query';
+import { AxiosResponse } from 'axios';
+
+export type InternalMutation<T> = UseMutationResult<
+	AxiosResponse<any, any>,
+	Error,
+	T,
+	unknown
+>;
